% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSignal.R
\name{addSignal}
\alias{addSignal}
\title{Loads NIfTI Images and Adds Synthetic Signal}
\usage{
addSignal(files, betaimg, X, Xred, outfiles = NULL, standardize = FALSE)
}
\arguments{
\item{files}{a vector of .nii or .nii.gz images.}

\item{betaimg}{a parameter image that describes the association between X
and the images in filelist. The units of this are standardized so that it
is like effect size.}

\item{X}{the design matrix for all covariates included in the simulation
model fit.}

\item{Xred}{the design matrix for all covariates except the column that is
being multiplied by betaimg. Xred must have only one less column than X.}

\item{outfiles}{a vector of images to save the output.}

\item{standardize}{Roughly standardize X and Y, so that the signal is homogenous across voxels and equal to effect size (roughly).}
}
\value{
Returns a 4d array of imaging data with synthetic signal added. The first three dimensions are equal to dim(betaimg) and the 4th dimension indexes subject.
}
\description{
This function loads images specified in filelist and adds signal to the
 images. The amount and structure of signal is determined by betaimg and X.
 The output images are equal the images specified by filelist plus betaimg
 times the column of X that is not in Xred. This column is first
 residualized to Xred.
}
\keyword{power}
\keyword{simulation}
