% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmap.R
\name{coef.statMap}
\alias{coef.statMap}
\title{Gets a 4D niftiImage of the coefficient image from a statMap object}
\usage{
\method{coef}{statMap}(x)
}
\arguments{
\item{x}{the statMap object to extract a coefficient niftiImage from}
}
\value{
a niftiImage object of the coefficient image
}
\description{
Given a statMap object and a directory write the objects as stat.nii.gz, sqrtSigma.nii.gz and summary.txt
}
