% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmap.R
\name{colorBar}
\alias{colorBar}
\title{Draws a color bar in a figure}
\usage{
colorBar(
  lut,
  min,
  max = -min,
  nticks = 4,
  ticks = seq(min, max, len = nticks),
  title = "",
  ylab = "",
  las = 1
)
}
\arguments{
\item{lut}{the range of colors to display}

\item{min}{Where to put the minimum tick mark.}

\item{max}{Where to put the maximum tick mark.}

\item{nticks}{Number of ticks on the color bar.}

\item{ticks}{Locations for ticks.}

\item{title}{The title for the color bar}

\item{ylab}{label for the color bar.}

\item{las}{argument passed to par.}
}
\value{
a niftiImage object of the coefficient image
}
\description{
Can be used in combination with layout and image.statMap to create figures.
}
\details{
Returns a statistical niftiImage object from a statMap object
}
