% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStats.R
\name{computeStats}
\alias{computeStats}
\title{Computes Statistical Map for Neuroimaging Data}
\usage{
computeStats(images, form, formred, mask, data = NULL, W = NULL,
  Winv = NULL, template = NULL, formImages = NULL, robust = TRUE,
  sqrtSigma = TRUE, transform = TRUE, outdir = NULL,
  mc.cores = getOption("mc.cores", 2L))
}
\arguments{
\item{images}{Character vector of subject images to be modeled as an outcome
variable OR 4d array of imaging data OR 4d nifti object.}

\item{form}{formula or character that can be coerced to a formula or a design
matrix for the full model.}

\item{formred}{formula or character that can be coerced to a formula or a
design matrix for reduced model. If robust=TRUE then this must have one less
column than X.}

\item{mask}{File name for a binary mask file or niftiImage object.}

\item{data}{R data frame containing variables in form. If form and formred
are matrices then this can be NULL.}

\item{W}{Weights for regression model. Can be used to deweight noisy
observations. Same as what should be passed to lm.}

\item{Winv}{Inverse weights for regression model. Inverse of W. Required when
passing variance images as the inverse weights.}

\item{template}{Template image used for visualization.}

\item{formImages}{n X p matrix of images where n is the number of subjects and
each column corresponds to an imaging covariate. Currently, not supported.}

\item{robust}{Logical, compute robust standard error estimates? Defaults to TRUE.
Uses HC3 SE estimates from Long and Ervin 2000.}

\item{sqrtSigma}{Logical: return V X n matrix sqrtSigma? Defaults to TRUE (described below).
Required to use pbj function.}

\item{transform}{Logical: use transformation from equation (5) of Vandekar et al. 2019 (Biometrics)?
Defaults to TRUE.
(instead of niftiImage objects; defaults to FALSE).}

\item{outdir}{If specified, output is saved as NIfTIs and statMap object is
saved as strings. This approach conserves memory, but has longer IO time.
Currently, not supported.}

\item{mc.cores}{Argument passed to mclapply for parallel things.}
}
\value{
Returns a list with the following values:
\describe{
  \item{stat}{The statistical values where mask!=0. If ncol(X) = ncol(Xred)+1, then this is a Z-statistic map, otherwise it is a chi^2-statistic map.}
  \item{coef}{A 4d niftiImage giving the parameter estimates for covariates only in the full model.}
  \item{sqrtSigma}{The 4d covariance object. This is a V by n matrix R, such that R \%*\% t(R) = hatSigma.}
  \item{mask}{The input mask.}
  \item{template}{The background template used for visualization.}
  \item{formulas}{A list containing the full and reduced models.}
  \item{robust}{A logical indicating the input setting.}
  \item{df}{The numerator degrees of freedom of the test.}
  \item{rdf}{The numerator degrees of freedom of the test.}
}
stat=stat, sqrtSigma=res, mask=mask, template=template, formulas=list(form, formred), robust=robust, df=df, rdf=rdf
}
\description{
This function computes a statistical map and residuals which are the
 objects necessary to perform the parametric bootstrap joint (PBJ) inference
 procedure.
}
\keyword{bootstrap}
\keyword{bootstrap,}
\keyword{map,}
\keyword{parametric}
\keyword{semiparametric}
\keyword{statistical}
