% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftiImage.R
\name{crop}
\alias{crop}
\title{Crop niftiImages}
\usage{
crop(mask, img)
}
\arguments{
\item{mask}{a mask image to crop with}

\item{img}{another image to crop}
}
\value{
A list with the cropped mask and img, and the offsets to adjust image indexing
}
\description{
Crop niftiImages
}
