% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalues.R
\name{getPvalues}
\alias{getPvalues}
\title{Gets p-values for Simulation Study}
\usage{
getPvalues(pmap, betamap = NULL, kernel = "box", pfunc = function(x)
  10^(-abs(x)))
}
\arguments{
\item{pmap}{A cluster-extent p-value map (nii, nii.gz, or Nifti object),
where each cluster is labeled with its FWE adjusted cluster-extent p-value.
(Required argument)}

\item{betamap}{A known betamap (nii, nii.gz, or Nifti object) used in simulations to generate artificial
signal in the imaging outcome. If NULL then power statistics are not
reported.}

\item{kernel}{What type of kernel to use for connected components. Defaults
to "box".}

\item{pfunc}{If pmap is transformed p-values then what function should be
used to compute untransformed p-values. Uses pbj default, which stores p-
values as -log10(p).}
}
\value{
Returns a list with the following values:
\item{altp}{If betamap is non-NULL then this returns a list the length of
 the number of connected components in betamap. Each list element gives
 the ordered p-values of clusters in pmap that have nonempty overlap with
 that cluster in betamap. The output is ordered by increasing size of
 connected components in betamap. NULL if betamap is NULL.}
\item{nullp}{If betamap is non-NULL then these are the p-values for clusters
 in pmap that have empty intersection with betamap. Otherwise, this is a
 vector of p-values for all clusters in pmap.}
\item{clustsize}{If betamap us non-NULL then this is the sizes of clusters in betamap in increasing order. NULL otherwise.}
\item{altinds}{If betamap is non-NULL then this returns a list the length of
 the number of connected components in betamap. Each list element gives the
 indices of clusters indices in componentmap that overlap with the
 corresponding cluster in betamap. NULL otherwise.}
\item{nullinds}{Vector of indices of components in component map that do not
 overlap with betamap.}
\item{componentmap}{Nifti object giving connected components of pmap.}
}
\description{
Gets p-values for Simulation Study
}
