% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbj.R
\name{image.CoPE}
\alias{image.CoPE}
\title{Image a CoPE object}
\usage{
\method{image}{CoPE}(x, alpha = 0.05, ...)
}
\arguments{
\item{x}{pbj object to create images for}

\item{alpha}{Displays 1-alpha CoPE maps}

\item{...}{Arguments passed to image.statMap}
}
\description{
See image.statMap for additional arguments
}
