% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmap.R
\name{image.statMap}
\alias{image.statMap}
\title{Create images of a statMap}
\usage{
\method{image}{statMap}(x, thresh = 2.32, index = NULL,
  col = gray(0:64/64), colpos = redyellow(64), colneg = bluecyan(64),
  plane = c("axial", "coronal", "sagittal"), xlab = "", ylab = "",
  axes = FALSE, oma = rep(0, 4), mar = rep(0, 4), bg = "black",
  ...)
}
\description{
modified from oro.nifti:::image.nifti
}
