% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbj.R
\name{inferenceIndex}
\alias{inferenceIndex}
\title{Gets numeric index of observed stat corresponding to the method and CFT}
\usage{
inferenceIndex(obsStat, method, cft = NULL)
}
\arguments{
\item{obsStat}{Field in statMap$pbj$obsStat}

\item{method}{What statistic to provide summary for? must have run that
analysis using the pbjInference and mmeStat functions.}

\item{cft}{cluster forming threshold to display. If NULL, just display the first.}
}
\value{
Returns an index that corresponds to the method and CFT.
}
\description{
Gets numeric index of observed stat corresponding to the method and CFT
}
\seealso{
[mmeStat], [cluster], [maxima], [pbjInference]
}
