% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npbj.R
\name{npbj}
\alias{npbj}
\title{Performs nonParametric Bootstrap Joint (nPBJ) Inference}
\usage{
npbj(images, form, formred, mask, data = NULL, W = NULL,
  template = NULL, nboot = 1000, statistic = sei, ...)
}
\arguments{
\item{images}{A character vector of subjects' images or a 4D niftiImage object.}

\item{form}{A formula describing the full model.}

\item{formred}{A formula describing the reduced model.}

\item{mask}{A character string of niftiImage object identifying voxels to be included in the analysis.}

\item{data}{A data frame with the variables in form for associated with subject images.}

\item{W}{A weight vector for the weighted least squares regression.}

\item{template}{A template image that will be included in the returned statMap object.}

\item{nboot}{Number of bootstraps to perform.}

\item{statistic}{A function of the form statistic(stat, ...) that takes a vector stat and other arguments
to compute an image summary statistic for each bootstrap. E.g. \code{\link{npbj.sei}}.}

\item{...}{Arguments passed to statistic function.}
}
\value{
Returns a list of length length(cfts)+2. The first two elements contain
statMap$stat and statMap$template. The remaining elements are lists containing the following:
\item{pvalues}{A vector of p-values corresponding to the cluster labels in clustermaps.}
\item{clustermap}{A niftiImage object with the cluster labels.}
\item{pmap}{A nifti object with each cluster assigned the negative log10 of its cluster extent FWE adjusted p-value.}
\item{CDF}{A bootstrap CDF.}
}
\description{
Performs nonParametric Bootstrap Joint (nPBJ) Inference
}
\seealso{
\code{\link{sei}}
}
