% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm.R
\name{palm}
\alias{palm}
\title{this is an R wrapper function for PALM
it has limited functionality for the features that PALM offers}
\usage{
palm(i, m, s = NULL, d, o, C = NULL, Cstat = "extent", tc = NULL,
  f = NULL, fonly = TRUE, n = 10000, eb = NULL, within = NULL,
  whole = NULL, ee = TRUE, T = FALSE, noniiclass = TRUE,
  logp = TRUE)
}
\arguments{
\item{i}{vector of strings pointing to input images or a single 4d nifti string.}

\item{m}{mask file.}

\item{s}{vector of strings pointing to input surfaces or a single 4d surface.}

\item{d}{design matrix as a data frame or csv file.}

\item{o}{output directory and prefix.}

\item{C}{cluster forming threshold as a Z-value}

\item{Cstat}{cluster statistic}

\item{tc}{T-contrast matrix that randomise takes as matrix, data frame, or csv.}

\item{f}{F-contrast matrix that randomise takes, as matrix data frame, or csv.}

\item{fonly}{only run the F-tests.}

\item{n}{number of permutations to run.}

\item{eb}{exchangeability blocks (not functional).}

\item{within}{not functional.}

\item{whole}{not functional.}

\item{ee}{exchangeable errors assumed.}

\item{T}{Run TFCE.}

\item{noniiclass}{Allow gzipped niftis.}

\item{logp}{Compute log10p-values.}
}
\value{
Returns a string of the command that was evaluated.
}
\description{
this is an R wrapper function for PALM
it has limited functionality for the features that PALM offers
}
