% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbjBoot.R
\name{pbjBoot}
\alias{pbjBoot}
\title{Performs (semi)Parametric Bootstrap Joint ((s)PBJ) Inference}
\usage{
pbjBoot(
  sqrtSigma,
  rboot,
  bootdim,
  method = c("nonparametric", "t", "conditional", "permutation"),
  voxelwise = FALSE,
  HC3 = TRUE,
  robust = TRUE,
  transform = c("none", "t")
)
}
\arguments{
\item{sqrtSigma}{list from statmap object}

\item{rboot}{a function that draws a bootstrapped sample. Can be an n vector, V by n matrix.}

\item{bootdim}{Dimension of rboot output}

\item{method}{Method to use for resampling.}

\item{voxelwise}{logical indicating whether the data are voxelwise.}

\item{HC3}{logical, was HC3 estimator used?}

\item{robust}{Generate robust statistics?}

\item{transform}{Apply a quantile transformation to the test statistics to improve normal approximation.}
}
\value{
Returns vector of test statistics computed from the bootstrapped sample.
}
\description{
Performs (semi)Parametric Bootstrap Joint ((s)PBJ) Inference
}
