% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbjClust.R
\name{pbjClust}
\alias{pbjClust}
\title{Performs (semi)Parametric Bootstrap Joint Inference ((s)PBJ)}
\usage{
pbjClust(stat = NULL, res = NULL, mask = NULL, statoutfiles = NULL,
  df = 0, rdf = NULL, cfts = c(0.01, 0.005), nboot = 5000,
  kernel = "box")
}
\arguments{
\item{stat}{staterved statistical map e.g. as obtained from computeStats.}

\item{res}{Character of 4d residuals in nii or nii.gz format, or residual
matrix as obtained from computeStats.}

\item{mask}{Character mask file location.}

\item{statoutfiles}{Character basename for nii.gz files to write out
-log10(p) maps. These are thresholded maps, where the value in each cluster
is the cluster adjusted p-value.}

\item{df}{Degrees of freedom of test statistics. This is the numerator
degrees of freedom of the F-statistic. If you are passing a Z-statistic
image set df=0.}

\item{rdf}{Residual degrees of freedom. This is the denominator degrees of
freedom of the F-statistic. This is an optional parameter. If rdf<<n then
this might save time, otherwise, it is ok to leave NULL.}

\item{cfts}{Numeric vector of cluster forming thresholds to use. These are
single-tailed for chi-squared (or F) statistics or two tailed for
Z-statistics probabilities.}

\item{nboot}{Number of bootstraps to perform.}

\item{kernel}{Kernel to use for computing connected components. Box is
default, but Diamond may also be reasonable.}
}
\description{
Performs (semi)Parametric Bootstrap Joint Inference ((s)PBJ)
}
