% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbjInference.R
\name{pbjInferenceBG}
\alias{pbjInferenceBG}
\title{Performs (semi)Parametric Bootstrap Joint ((s)PBJ) Inference}
\usage{
pbjInferenceBG(argsList)
}
\arguments{
\item{argsList}{A list of arguments obtained from pbjInference including `rdata_rds` that specifies a file to save the output.}
}
\description{
Performs (semi)Parametric Bootstrap Joint ((s)PBJ) Inference
}
\details{
This function performs pbj resampling-based inference running the analyses in the background and saving the results as and rdata or rds object.
If the file name ends in rdata it will save the statMap object and the computing time in an rdata file. Otherwise it will save the statMap object as an rds file.
Note, the input to this function is different than [pbjInferenceFG()].
}
\seealso{
[pbjInference()], [pbjInferenceFG()]
}
