% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbjSEI.R
\name{pbjSEI}
\alias{pbjSEI}
\title{Performs (semi)Parametric Bootstrap Joint ((s)PBJ) Spatial Extent Inference}
\usage{
pbjSEI(statMap, cfts.s = c(0.1, 0.25), cfts.p = NULL, nboot = 5000,
  kernel = "box", rboot = stats::rnorm)
}
\arguments{
\item{statMap}{statMap object as obtained from computeStats.}

\item{cfts.s}{Numeric vector of cluster forming thresholds to use based on the robust effect size index.}

\item{cfts.p}{Numeric vector of clusterforming thresholds to use based on p-value thresholding. Consistent
with other imaging software these thresholds are one tailed (e.g. p<0.01 imples z>2.32).}

\item{nboot}{Number of bootstrap samples to use.}

\item{kernel}{Kernel to use for computing connected components. box is
default (26 neighbors), but diamond may also be reasonable. argument to mmand::shapeKernel}

\item{rboot}{Function for generating random variables. See examples.}
}
\value{
Returns a list of length length(cfts)+4. The first four elements contain
statMap$stat, statMap$template, statMap$mask, and statMap$df. The remaining elements are lists containing the following:
\item{pvalues}{A vector of p-values corresponding to the cluster labels in clustermaps.}
\item{clustermap}{A niftiImage object with the cluster labels.}
\item{pmap}{A nifti object with each cluster assigned the negative log10 of its cluster extent FWE adjusted p-value.}
\item{CDF}{A bootstrap CDF.}.
}
\description{
Performs (semi)Parametric Bootstrap Joint ((s)PBJ) Spatial Extent Inference
}
