% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmap.R
\name{plotData.statMap}
\alias{plotData.statMap}
\title{Creates dataframes, emmeans, and models of the selected ROI data}
\usage{
plotData.statMap(
  x,
  emForm = NULL,
  method = "CEI",
  cft_s = NULL,
  cft_p = NULL,
  cft_chisq = NULL,
  roiInds = NULL,
  ...
)
}
\arguments{
\item{x}{the statMap object}

\item{emForm}{a formula specifying how to plot the data}

\item{method}{A character specifying which inference method to plot results for. One of maxima, CEI, or CMI}

\item{cft_s}{cluster forming threshold on effect size scale.}

\item{cft_p}{cluster forming threshold on p-value scale.}

\item{cft_chisq}{cluster forming threshold on chi-square statistic scale.}

\item{roiInds}{A numeric/integer vector specifying which ROIs to plot results for.}

\item{...}{arguments passed to ref_grid}
}
\value{
a list with a dataframe containing the original data with a column added for
mean ROI values for each subject, an emmeans objects based on the emForm argument,
and model objects of the full model fit to each ROI.
}
\description{
Returns a statistical niftiImage object from a statMap object
}
\details{
Returns ROI summary results if pbjInference has been executed on the statMap object.
If not, it attempts to create a cluster summary table based on the method and cft_* arguments.
emmeans objects can be used for plotting and are returned if emForm is specified.
}
