% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmap.R
\name{ptfce.statmap}
\alias{ptfce.statmap}
\title{Compute pTFCE image}
\usage{
ptfce.statmap(statMap, ...)
}
\arguments{
\item{statMap}{A statmap object.}

\item{...}{Arguments passed to ptfce.}
}
\description{
Uses a statMap object to perform pTFCE inference
}
