% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNiftis.R
\name{readNiftis}
\alias{readNiftis}
\title{A Function to read NIfTI files in parallel}
\usage{
readNiftis(files = NULL, mc.cores = getOption("mc.cores", 2L))
}
\arguments{
\item{files}{A vector of image names.}

\item{mc.cores}{number of cores to use when loading images}
}
\description{
This function is a wrapper for readNifti that opens a vector of nii or
nii.gz files in parallel.
}
\keyword{cats}
