% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbj.R
\name{resics}
\alias{resics}
\title{Performs one bootstrap for RESI confidence set construction}
\usage{
resics(stat, ShatSq, n, df, rdf, normMethod = c("1", "param"), mask)
}
\arguments{
\item{stat}{bootstrap value of the test statistic image. It is a nifti image array.}

\item{ShatSq}{The RESI estimate computed in the observed data.}

\item{n}{The sample size/ number of independent sampling units.}

\item{df}{Numerator degrees of freedom}

\item{rdf}{Denominator degrees of freedom}

\item{normMethod}{The value used to normalize the test statistic in each
bootstrap. "param" means it's estimated assuming the test
statistic if F-distributed. "1" means no scaling. "1" is the default.}

\item{mask}{Mask image identifying where measurements were taken in the image.}
}
\description{
Performs one bootstrap for RESI confidence set construction
}
