% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualizeImages.R
\name{residualizeImages}
\alias{residualizeImages}
\title{Prepares Simulation Data for Bootstrapping}
\usage{
residualizeImages(
  files,
  form,
  dat,
  mask,
  outfiles,
  smoutfiles = NULL,
  sm = 0,
  mc.cores = getOption("mc.cores", 2L)
)
}
\arguments{
\item{files}{Character vector of subject images to be modeled as an outcome
variable.}

\item{form}{mgcv or lm style formula.}

\item{dat}{Data frame containing covariates used by form.}

\item{mask}{Character giving location of mask image.}

\item{outfiles}{Character vector of residual output images to smooth.}

\item{smoutfiles}{Character vector of smoothed output images.}

\item{sm}{Numeric giving the smoothing amount in mm FWHM to perform before creating residuals. Smoothing is performed using fsl's susan.}

\item{mc.cores}{Argument passed to mclapply for parallel things.}
}
\value{
No returned value. This functions saves out nifti images files after residualizing to the model specified by form and dat. The residuals of files are saved as the corresponding element in outfiles.
}
\description{
Residualizes the images in files to the model form and writes the output to
outfiles optionally smoothes with sm (in mm FWHM) using susan prior to
residualizing if smoutfiles is specified smoothed images are saved into that
directory. Used to run simulations to assess power and type 1 error for
papers. This creates images that are residualized to the covariates which
can then be bootstrapped to generate a sample where there is the potential
for heteroskedasticity/nonexchangeability, but where the covariates are
unassociated with the mean of the outcome.
}
\keyword{1}
\keyword{bootstrap,}
\keyword{error}
\keyword{null}
\keyword{parametric}
\keyword{power}
\keyword{simulation,}
\keyword{simulations}
\keyword{simulations,}
\keyword{type}
