% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmap.R
\name{roiMeans}
\alias{roiMeans}
\title{Plots the tested variables in a statMap object}
\usage{
roiMeans(statMap, method = "CEI", cft = NULL, roiInds = NULL, data = NULL)
}
\arguments{
\item{statMap}{the statMap object with pbjInference run}

\item{method}{A character specifying which inference method to plot results for. One of maxima, CEI, or CMI}

\item{cft}{A numeric specifying CFT to plot results for, defaults to the first one}

\item{roiInds}{A numeric/integer vector specifying which ROIs to plot results for.}

\item{data}{Data frame with covariates. Optional if available in statMap object.}
}
\value{
a niftiImage object of the coefficient image
}
\description{
Returns a statistical niftiImage object from a statMap object
}
