% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbj.R
\name{sci}
\alias{sci}
\title{Uses bootstrap results to construct simultaneous confidence intervals.}
\usage{
sci(statmap, alpha)
}
\arguments{
\item{statmap}{The statMap object returned by lmPBJ using the `resics` function.}

\item{alpha}{The function returns 1-alpha simultaneous confidence intervals.}
}
\description{
Uses bootstrap results to construct simultaneous confidence intervals.
}
