% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmap.R
\name{summary.statMap}
\alias{summary.statMap}
\title{object A statMap object}
\usage{
\method{summary}{statMap}(object, ...)
}
\arguments{
\item{object}{A statMap object}

\item{...}{additional arguments}
}
\description{
object A statMap object
}
