% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbj.R
\name{table.pbj}
\alias{table.pbj}
\title{Create cluster/maxima summary table}
\usage{
table.pbj(x, method = c("CEI", "maxima", "CMI"))
}
\arguments{
\item{x}{pbj object}

\item{method}{What statistic to provide summary for? must have run that
analysis using the pbjInference and mmeStat functions.}
}
\value{
Returns table of unadjusted and FWER adjusted p-values and other summary statistics. Results depend on what statistic
function was used for pbjInference.
}
\description{
Create cluster/maxima summary table
}
\seealso{
[mmeStat], [pbjInference]
}
