% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmap.R
\name{table.statMap}
\alias{table.statMap}
\title{Create cluster/maxima summary table}
\usage{
table.statMap(
  x,
  method = c("CEI", "maxima", "CMI"),
  cft_s = NULL,
  cft_p = NULL
)
}
\arguments{
\item{x}{statMap object}

\item{method}{What statistic to provide summary for? must have run that
analysis using the pbjInference and mmeStat functions.}

\item{cft_s}{cluster forming threshold on effect size scale to show cluster p-values for.}

\item{cft_p}{cluster forming threshold on p-value scale to show cluster p-values for.}
}
\value{
Returns table of unadjusted and FWER adjusted p-values and other summary statistics. Results depend on what statistic
function was used for pbjInference.
}
\description{
Create cluster/maxima summary table
}
\details{
If both cft_s and cft_p are NULL, then it returns the first CFT that was used.
}
\seealso{
[mmeStat], [cluster], [maxima], [pbjInference]
}
