% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmap.R
\name{var.statMap}
\alias{var.statMap}
\title{Gets a niftiImage of the variance image from a statMap object}
\usage{
var.statMap(x)
}
\arguments{
\item{x}{the statMap to extract the variance image from}
}
\value{
a niftiImage object of the variance image
}
\description{
Will return a niftiImage of the variance image
}
