% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbj.R
\name{wecdf}
\alias{wecdf}
\title{Computes empirical weighted cdf. Modified from ecdf}
\usage{
wecdf(x, w = rep(1, length(x)))
}
\arguments{
\item{x}{vector of values}

\item{w}{vector with length(w)=length(x) of weights}
}
\value{
Returns list of tables of sizes of the connected components above thr.
}
\description{
Computes empirical weighted cdf. Modified from ecdf
}
