% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbj.R
\name{write.pbj}
\alias{write.pbj}
\title{Write a pbj object to disk}
\usage{
write.pbj(x, outdir, ...)
}
\arguments{
\item{x}{pbj object to write}

\item{outdir}{output directory to write pbj pieces}

\item{...}{additional arguments; unused.}
}
\description{
Write a pbj object to disk in parts
}
