% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootStats.R
\name{bootStats}
\alias{bootStats}
\title{Computes Statistical Map for Neuroimaging Data}
\usage{
bootStats(
  images,
  coefficients = NULL,
  X,
  Xred,
  statistic = function(stat, ...) stat,
  ...
)
}
\arguments{
\item{images}{An n by V  matrix of voxels inside the brain.}

\item{coefficients}{An n by V matrix of parameter estimates to subtract from the bootstrapped values.
If computing p-values this should be the observed parameter estimates. If computing confidence intervals
then this should be NULL (the default).}

\item{X}{Design matrix for the full model.}

\item{Xred}{Design matrix for the reduced model.}

\item{statistic}{A function of the form statistic(stat, ...) that returns a statistic of interest.}

\item{...}{Arguments passed to statistic function.}
}
\value{
Returns a vector of the resulting estimate
}
\description{
This function computes a statistical map and residuals which are the
 objects necessary to perform the parametric bootstrap joint (PBJ) inference
 procedure.
}
\keyword{bootstrap}
\keyword{map,}
\keyword{nonparametric}
\keyword{parametric}
\keyword{statistical}
