% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbj.R
\name{cluster}
\alias{cluster}
\title{Computes contiguous clusters from a statistical image given a threshold}
\usage{
cluster(
  stat,
  mask,
  thr,
  method = c("extent", "mass"),
  kernel = "box",
  rois = FALSE
)
}
\arguments{
\item{stat}{A statistical Nifti image as an RNifti image object.}

\item{mask}{A statistical Nifti image mask used in the analysis or a character path to one.}

\item{thr}{Vector of thresholds to threshold the test statistic image.}

\item{method}{character string 'extent' or 'mass' indicating whether the cluster extent or cluster mass statistic should be used.}

\item{kernel}{The kernel type to compute connected components.}

\item{rois}{If TRUE, return image with voxel values having the indices of the clusters returned if rois=FALSE.}
}
\value{
Returns list of tables of sizes of the connected components above thr.
}
\description{
Computes contiguous clusters from a statistical image given a threshold
}
