% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftiImage.R
\name{image.niftiImage}
\alias{image.niftiImage}
\title{Create images of a NiFTI object}
\usage{
\method{image}{niftiImage}(
  x,
  BGimg = NULL,
  limits = 0,
  nrow = NULL,
  index = NULL,
  crop = FALSE,
  col = gray(0:64/64),
  colpos = pbj:::redyellow(64),
  colneg = pbj:::bluecyan(64),
  plane = c("axial", "coronal", "sagittal"),
  title = "",
  axes = FALSE,
  lo = TRUE,
  other = function() {
 },
  ...
)
}
\arguments{
\item{x}{the Nifti object to display images of}

\item{BGimg}{background image to use.}

\item{limits}{A lower (and optionally upper length 2 vector) limitsold to apply to the image, defaults to 0}

\item{nrow}{number of rows to display the slices as. NULL is default and it's a square shape determined by length of `index` argument.}

\item{index}{Any selected image planes. defaults to NULL}

\item{crop}{crop white space from image. Default is FALSE.}

\item{col}{a vector of colors to use for scaled intensities defaults to a grey scale.}

\item{colpos}{a vector of colors to use for positive values.}

\item{colneg}{a vector of colors to use for negative values.}

\item{plane}{the plane to display, can be axial, coronal or sagittal.}

\item{title}{Title for figure drawn in outer margin with `mtext`.}

\item{axes}{display axes, defaults to false}

\item{lo}{Call layout to arrange images? Default is TRUE.}

\item{other}{A function used to add features to the plot.}

\item{...}{additional arguments passed to par}
}
\description{
modified from oro.nifti:::image.nifti
}
\examples{
library(RNifti)
library(pain21)
pain = pain21()
templ = readNifti(pain$template)
# image(templ)
# image(templ, plane='coronal')

}
