% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbj.R
\name{image.pbj}
\alias{image.pbj}
\title{Image a pbj object}
\usage{
\method{image}{pbj}(x, alpha = 0.05, ...)
}
\arguments{
\item{x}{pbj object to create images for}

\item{alpha}{numeric; threshold to apply for threshold mask of significant voxels}

\item{...}{Arguments passed to image.statMap}
}
\description{
See image.statMap for additional arguments
}
