% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmap.R
\name{image.statMap}
\alias{image.statMap}
\title{View statMap and pbj results in a statMap object}
\usage{
\method{image}{statMap}(
  x,
  method = c("CEI", "maxima", "CMI"),
  cft_s = NULL,
  cft_p = NULL,
  cft_chisq = NULL,
  roi = NULL,
  index = NULL,
  alpha = NULL,
  clusterMask = TRUE,
  clusterID = TRUE,
  title = "",
  plane = c("axial", "sagittal", "coronal"),
  crop = TRUE,
  lo = TRUE,
  ...
)
}
\arguments{
\item{x}{the statMap that created the pbj object.}

\item{method}{Which inference method to visualize.}

\item{cft_s}{The cluster forming threshold or threshold for visualizing results on the RESI effect size scale.}

\item{cft_p}{The cluster forming threshold or threshold for visualizing results on the p-value scale.}

\item{cft_chisq}{The cluster forming threshold or threshold for visualizing results on the chi-squared scale.}

\item{roi}{Which ROI index to visualize.}

\item{index}{Which slices to visualize. Default is all slices}

\item{alpha}{Adjusted p-value threshold to display clusters.}

\item{clusterMask}{logical indicating whether to mask the results with significant clusters.}

\item{clusterID}{logical indicating whether to include the clusterID in the figure.}

\item{title}{Character passed to image.niftiImage}

\item{plane}{Which anatomical plane to visualize.}

\item{crop}{Whether to crop white space from image. Defaults to TRUE.}

\item{lo}{Whether to perform layout call to arrange images. Defaults to true.}

\item{...}{Arguments passed to par.}
}
\description{
Uses a statMap with inferences results to visualize CEI, CMI, or maxima
}
