% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbj.R
\name{maxima}
\alias{maxima}
\title{Computes local maxima from an nifti image}
\usage{
maxima(stat, kernel = "box", width = 0, rois = FALSE, roisWidth = 15)
}
\arguments{
\item{stat}{A statistical Nifti image as an RNifti image object.}

\item{kernel}{Type of kernel to use for max/dilation filter}

\item{width}{Width of kernel (assumes isotropic). If zero, then returns global maximum.}

\item{rois}{If TRUE, return image with voxel values having the indices of the local maxima.}

\item{roisWidth}{Width of kernel for computing local maxima used for table output later. Only used if width is zero}
}
\value{
Default returns vector of local maxima in the image.
}
\description{
Computes local maxima from an nifti image
}
