% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbjBoot.R
\name{pbjBoot}
\alias{pbjBoot}
\title{Performs (semi)Parametric Bootstrap Joint ((s)PBJ) Inference}
\usage{
pbjBoot(
  sqrtSigma,
  rboot = function(n) {     (2 * stats::rbinom(n, size = 1, prob = 0.5) - 1) },
  bootdim,
  method = c("wild", "permutation", "nonparametric"),
  HC3 = TRUE,
  robust = TRUE,
  transform = c("none", "t")
)
}
\arguments{
\item{sqrtSigma}{list from statmap object}

\item{rboot}{a function that draws a bootstrapped sample. Should return an n vector. Defaults to Rademacher random variable.}

\item{bootdim}{Dimension of rboot output}

\item{method}{character, method to use for resampling procedure. Wild bootstrap, permutation, or nonparametric}

\item{HC3}{logical, was HC3 estimator used?}

\item{robust}{Generate robust statistics?}

\item{transform}{Apply a quantile transformation to the test statistics to improve normal approximation.}
}
\value{
Returns vector of test statistics computed from the bootstrapped sample.
}
\description{
Performs (semi)Parametric Bootstrap Joint ((s)PBJ) Inference
}
