% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbjBoot.R
\name{pbjBoot}
\alias{pbjBoot}
\title{Performs (semi)Parametric Bootstrap Joint ((s)PBJ) Inference}
\usage{
pbjBoot(
  sqrtSigma,
  rboot,
  bootdim,
  V,
  n,
  df,
  randomX = FALSE,
  robust = TRUE,
  method = c("nonparametric", "t", "conditional", "permutation", "robustPermutation"),
  voxelwise = FALSE,
  HC3 = TRUE
)
}
\arguments{
\item{sqrtSigma}{list from statmap object}

\item{rboot}{a function that draws a bootstrapped sample. Can be an n vector, V by n matrix.}

\item{bootdim}{Dimension of rboot output}

\item{V}{Number of voxels in analysis.}

\item{n}{Number of subjects.}

\item{df}{Numerator degrees of freedom.}

\item{randomX}{Generate X in each bootstrap as though it is random as well.}

\item{robust}{Generate robust statistics?}

\item{method}{Method to use for resampling.}

\item{voxelwise}{logical indicating whether the data are voxelwise.}

\item{HC3}{logical, use HC3 residual estimator?}
}
\value{
Returns vector of test statistics computed from the bootstrapped sample.
}
\description{
Performs (semi)Parametric Bootstrap Joint ((s)PBJ) Inference
}
