% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbjClust.R
\name{pbjClust}
\alias{pbjClust}
\title{Performs (semi)Parametric Bootstrap Joint ((s)PBJ) Spatial Extent Inference}
\usage{
pbjClust(statMap, cfts = c(0.01, 0.005), df = 0, rdf = NULL,
  nboot = 5000, kernel = "box")
}
\arguments{
\item{statMap}{statMap object as obtained from computeStats.}

\item{cfts}{Numeric vector of cluster forming thresholds to use. These are
single-tailed for chi-squared (or F) statistics or two tailed for
Z-statistics probabilities.}

\item{df}{Degrees of freedom of test statistics. This is the numerator
degrees of freedom of the F-statistic. If you are passing a Z-statistic
image set df=0.}

\item{rdf}{Residual degrees of freedom. This is the denominator degrees of
freedom of the F-statistic. This is an optional parameter. If rdf<<n then
this might save time, otherwise, it is ok to leave NULL.}

\item{nboot}{Number of bootstraps to perform.}

\item{kernel}{Kernel to use for computing connected components. Box is
default (26 neighbors), but Diamond may also be reasonable.}
}
\value{
Returns a list of length length(cfts)+2. The first two elements contain
statMap$stat and statMap$template. The remaining elements are lists containing the following:
\item{pvalues}{A vector of p-values corresponding to the cluster labels in clustermaps.}
\item{clustermap}{A niftiImage object with the cluster labels.}
\item{pmap}{A nifti object with each cluster assigned the negative log10 of its cluster extent FWE adjusted p-value.}
\item{CDF}{A bootstrap CDF.}
}
\description{
Performs (semi)Parametric Bootstrap Joint ((s)PBJ) Spatial Extent Inference
}
