% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbjClust.R
\name{pbjClust}
\alias{pbjClust}
\title{Performs (semi)Parametric Bootstrap Joint ((s)PBJ) Spatial Extent Inference}
\usage{
pbjClust(statMap, cfts = c(0.01, 0.005), nboot = 5000,
  kernel = "box")
}
\arguments{
\item{statMap}{statMap object as obtained from computeStats.}

\item{cfts}{Numeric vector of cluster forming thresholds to use. Consistent
with other imaging software these thresholds are one tailed (e.g. p<0.01 imples z>2.32).}

\item{nboot}{Number of bootstraps to perform.}

\item{kernel}{Kernel to use for computing connected components. Box is
default (26 neighbors), but Diamond may also be reasonable.}
}
\value{
Returns a list of length length(cfts)+4. The first four elements contain
statMap$stat, statMap$template, statMap$mask, and statMap$df. The remaining elements are lists containing the following:
\item{pvalues}{A vector of p-values corresponding to the cluster labels in clustermaps.}
\item{clustermap}{A niftiImage object with the cluster labels.}
\item{pmap}{A nifti object with each cluster assigned the negative log10 of its cluster extent FWE adjusted p-value.}
\item{CDF}{A bootstrap CDF.}
}
\description{
Performs (semi)Parametric Bootstrap Joint ((s)PBJ) Spatial Extent Inference
}
