% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbjExSet.R
\name{pbjExSet}
\alias{pbjExSet}
\title{Builds (semi)Parametric Bootstrap Joint ((s)PBJ) Coverage Probability Excursion (CoPE) Sets}
\usage{
pbjExSet(statMap, rsq = 0.05, nboot = 5000, boundary = FALSE,
  eps = 0.01)
}
\arguments{
\item{statMap}{statMap object as obtained from computeStats.}

\item{rsq}{minimum effect size of interest. Currently, only supported for length(rsq)=1.}

\item{nboot}{Number of bootstraps to perform.}

\item{boundary}{Sample from the boundary as in Sommerfeld et al. (2018).}
}
\value{
Returns a list of length length(rsq)+2. The first two elements contain
statMap$stat and statMap$template. The remaining elements are lists containing the following:
\item{Aminus}{Aminus<=1-alpha contains E{statMap$stat}>chsq with probability 1-alpha.}
\item{Aplus}{Aplus>1-alpha is contained in E{statMap$stat}>chsq with probability 1-alpha.}
\item{rsq}{A minimum threshold to define interesting locations. It is the proportion of signal in the chi-square statistic.}
\item{chsq}{rsq converted to the chi-squared scale using chsq = ( (n-1) \* rsq + 1) /(1-rsq) \* df.}
\item{CDFs}{The bootstrap CDFs used to compute Aminus and Aplus.}
}
\description{
Builds (semi)Parametric Bootstrap Joint ((s)PBJ) Coverage Probability Excursion (CoPE) Sets
}
