% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbjInference.R
\name{pbjInference}
\alias{pbjInference}
\title{Performs (semi)Parametric Bootstrap Joint ((s)PBJ) Inference}
\usage{
pbjInference(
  statMap,
  statistic = function(image) max(c(image)),
  randomX = FALSE,
  nboot = 5000,
  rboot = stats::rnorm,
  method = c("t", "condtional", "permutation"),
  ...
)
}
\arguments{
\item{statMap}{statMap object as obtained from computeStats.}

\item{statistic}{A user specified function that takes a RNift image object and computes a particular statistic of interest.}

\item{randomX}{logical, bootstrap samples of X as well?}

\item{nboot}{Number of bootstrap samples to use.}

\item{rboot}{Function for generating random variables. See examples.}

\item{method}{character method to use for bootstrap procedure.}

\item{...}{arguments passed to statistic function.}
}
\value{
Returns a list of length 2. The first element is the observed statistic value and the second is a list of the boostrap values.
}
\description{
Performs (semi)Parametric Bootstrap Joint ((s)PBJ) Inference
}
