% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbjInference.R
\name{pbjInference}
\alias{pbjInference}
\title{Performs (semi)Parametric Bootstrap Joint ((s)PBJ) Inference}
\usage{
pbjInference(
  statMap,
  statistic = function(image) max(c(image)),
  nboot = 5000,
  rboot = function(n) {     (2 * stats::rbinom(n, size = 1, prob = 0.5) - 1) },
  method = c("wild", "permutation", "nonparametric"),
  runMode = c("bootstrap", "cdf"),
  ...
)
}
\arguments{
\item{statMap}{statMap object as obtained from computeStats.}

\item{statistic}{A user specified function that takes a RNifti image object and computes a particular statistic of interest.}

\item{nboot}{Number of bootstrap samples to use.}

\item{rboot}{Function for generating random variables. Should return an n vector. Defaults to Rademacher random variable.}

\item{method}{Character, method to use for resampling procedure. Wild bootstrap, permutation, or nonparametric}

\item{runMode}{character, that controls output. cdf returns the empirical CDFs, bootstrap returns the bootstrapped statistics as a list.}

\item{...}{arguments passed to statistic function.}
}
\value{
Returns a list. if runMode=='bootstrap', the first element is the observed statistic value and the second is a list of the boostrap values. If runMode=='cdf', the first element is the observed statistic value, and the subsequent elements are the CDFs and ROIs, used for computing adjusted p-values and plotting.
}
\description{
Performs (semi)Parametric Bootstrap Joint ((s)PBJ) Inference
}
