% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmap.R
\name{plot.statMap}
\alias{plot.statMap}
\title{Plots the tested variables in a statMap object}
\usage{
\method{plot}{statMap}(
  x,
  emForm = NULL,
  method = "CEI",
  cft_s = NULL,
  cft_p = NULL,
  cft_chisq = NULL,
  roiInds = NULL,
  ...
)
}
\arguments{
\item{x}{the statMap object with pbjInference run}

\item{emForm}{a formula specifying how to plot the data}

\item{method}{A character specifying which inference method to plot results for. One of maxima, CEI, or CMI}

\item{cft_s}{cluster forming threshold on effect size scale to show cluster p-values for.}

\item{cft_p}{cluster forming threshold on p-value scale to show cluster p-values for.}

\item{cft_chisq}{cluster forming threshold on chi-square statistic scale to show cluster p-values for.}

\item{roiInds}{A numeric/integer vector specifying which ROIs to plot results for.}

\item{...}{arguments passed to plot}
}
\value{
a niftiImage object of the coefficient image
}
\description{
Returns a statistical niftiImage object from a statMap object
}
