% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmap.R
\name{plot.statMap}
\alias{plot.statMap}
\title{Plots the tested variables in a statMap object}
\usage{
\method{plot}{statMap}(statMap, emForm = NULL, method = "CEI", cft = NULL, roiInds = NULL)
}
\arguments{
\item{statMap}{the statMap object with pbjInference run}

\item{emForm}{a formula specifying how to}

\item{method}{A character specifying which inference method to plot results for. One of maxima, CEI, or CMI}

\item{cft}{A numeric specifying CFT to plot results for, defaults to the first one}

\item{roiInds}{A numeric/integer vector specifying which ROIs to plot results for.}
}
\value{
a niftiImage object of the coefficient image
}
\description{
Returns a statistical niftiImage object from a statMap object
}
