% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmap.R
\name{stat.statMap}
\alias{stat.statMap}
\title{Gets a 4D niftiImage of the coefficient image from a statMap object}
\usage{
stat.statMap(x)
}
\arguments{
\item{x}{the statMap object to extract a coefficient niftiImage from}
}
\value{
a niftiImage object of the coefficient image
}
\description{
Returns a statistical niftiImage object from a statMap object
}
