% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbj.R
\name{table.statMap}
\alias{table.statMap}
\title{Create cluster/maxima summary table}
\usage{
table.statMap(x, method = c("CEI", "maxima", "CMI"), cft = NULL)
}
\arguments{
\item{x}{pbj object}

\item{method}{What statistic to provide summary for? must have run that
analysis using the pbjInference and mmeStat functions.}

\item{cft}{cluster forming threshold to display. If NULL, just display the first.}
}
\value{
Returns table of unadjusted and FWER adjusted p-values and other summary statistics. Results depend on what statistic
function was used for pbjInference.
}
\description{
Create cluster/maxima summary table
}
\seealso{
[mmeStat], [cluster], [maxima], [pbjInference]
}
