% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbj.R
\name{vpapx_edgeworth}
\alias{vpapx_edgeworth}
\title{A vectorized version of papx_edgeworth}
\usage{
vpapx_edgeworth(stat, mu3, mu4)
}
\arguments{
\item{stat}{Vector of test statistics}

\item{skew}{The third cumulant of the test statistics}
}
\description{
See image.statMap for additional arguments
}
