% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmap.R
\name{write.statMap}
\alias{write.statMap}
\title{Write the statMap objects out}
\usage{
write.statMap(
  x,
  outdir,
  images = TRUE,
  sqrtSigma = TRUE,
  statMethod = c("p", "S", "chisq")
)
}
\arguments{
\item{x}{the statMap object to write out}

\item{outdir}{the directory to write into}

\item{images}{Logical indicating whether or not to write out the nifti images.}

\item{sqrtSigma}{Logical indicating whether or not to write out the objects needed for bootstrapping.}

\item{statMethod}{Passed to stat.statMap to specify how to write the statistical image. p= -log10(p) * sign(coef), S=RESI, chisq=chi-squared statistic}
}
\value{
a list of what was written
}
\description{
Given a statMap object and a directory write the objects as stat.nii.gz, coef.nii.gz and sqrtSigma.nii.gz
}
